[encointer-balances] adjust unit of `FeeConversionFactor` to be able to be adjusted more freely

+++ primitives/src/balances.rs
@@ -41,7 +41,7 @@ pub type BalanceType = I64F64;
 /// 1.3188e-07 : halving time of 1 year if blocktime is 6s
 pub type Demurrage = I64F64;
 
-pub type FeeConversionFactorType = u32;
+pub type FeeConversionFactorType = u128;
 
 #[derive(
 	Encode, Decode, Default, RuntimeDebug, Clone, Copy, PartialEq, TypeInfo, MaxEncodedLen,
